+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - AI Guard: `LlmInfo` and `LlmInput`.
+- PangeaConfig: `ConfigID` and `Enviroment` field.
 
 ## 4.4.0 - 2025-02-16